requirements.txt
google-cloud-storage
google-cloud-logging
paramiko
functions-framework